pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "langgraph-cua"
version = "0.0.1"
description = "An implementation of a computer use agent (CUA) using LangGraph"
authors = [
    {name = "Brace Sproul", email = "46789226+bracesproul@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.3.17,<0.4.0",
    "langchain-core>=0.3.46,<0.4.0",
    "scrapybara>=2.4.1,<3.0.0",
    "langchain-openai>=0.3.10,<0.4.0",
    "hyperbrowser>=0.38.0",
    "playwright>=1.51.0",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "ruff>=0.9.4",
    "mypy>=1.8.0",
    "pytest-socket>=0.7.0",
    "types-setuptools>=69.0.0",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q -v"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = [
  "E501" # line-length
]


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
